# backend/src/main/resources/openapi/registry-api.yaml
openapi: 3.0.3
info:
  title: HeatMap Registry API
  description: Реестр AI-инструментов PDLC
  version: 0.1.0
servers:
  - url: /
...
security: []
paths:
...
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Текущий пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Auth]
      operationId: updateMe
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          $ref: '#/components/responses/Error'

  # ===================== Tools =====================
...
  /api/admin/tools/pending:
    get:
      tags: [Admin]
      operationId: listPendingTools
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Заявки на модерации
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/{id}/approve:
    post:
      tags: [Admin]
      operationId: approveTool
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      responses:
        '200':
          description: Заявка одобрена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/{id}/reject:
    post:
      tags: [Admin]
      operationId: rejectTool
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectToolRequest'
      responses:
        '200':
          description: Заявка отклонена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/archived:
    get:
      tags: [Admin]
      operationId: listArchivedTools
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Архивные инструменты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/{id}/archive:
    post:
      tags: [Admin]
      operationId: archiveTool
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveToolRequest'
      responses:
        '200':
          description: Инструмент архивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/{id}/restore:
    post:
      tags: [Admin]
      operationId: restoreTool
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ToolId'
      responses:
        '200':
          description: Инструмент восстановлен из архива
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
        default:
          $ref: '#/components/responses/Error'

  # ===================== Admin: notes =====================
  /api/admin/tools/{toolId}/notes:
    parameters:
      - name: toolId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Admin]
      operationId: listToolNotes
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Заметки к инструменту
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ToolNoteResponse' }
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Admin]
      operationId: createToolNote
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolNoteRequest'
      responses:
        '200':
          description: Заметка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolNoteResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/tools/notes/{noteId}:
    parameters:
      - name: noteId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Admin]
      operationId: updateToolNote
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolNoteRequest'
      responses:
        '200':
          description: Заметка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolNoteResponse'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Admin]
      operationId: deleteToolNote
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Заметка удалена
        default:
          $ref: '#/components/responses/Error'

  # ===================== Admin: users =====================
  /api/admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/admin/users/{id}/role:
    patch:
      tags: [Admin]
      operationId: updateUserRole
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: Роль обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/users/{id}/enabled:
    patch:
      tags: [Admin]
      operationId: updateUserEnabled
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserEnabledRequest'
      responses:
        '200':
          description: Доступ пользователя обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/admin/users/{id}:
    delete:
      tags: [Admin]
      operationId: deleteUser
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Пользователь удалён
        default:
          $ref: '#/components/responses/Error'

  # ===================== Admin: impact =====================
  /api/admin/impact/rows/{id}:
    patch:
      tags: [Admin]
      operationId: updateImpactRow
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateImpactRowRequest'
      responses:
        '200':
          description: Строка метрики обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactRowResponse'
        default:
          $ref: '#/components/responses/Error'

  # ===================== Impact (public) =====================
...
  /api/notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Уведомления пользователя
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/NotificationResponse' }
        default:
          $ref: '#/components/responses/Error'
  /api/notifications/unread-count:
    get:
      tags: [Notifications]
      operationId: getUnreadNotificationCount
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Количество непрочитанных уведомлений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/notifications/{id}/read:
    post:
      tags: [Notifications]
      operationId: markNotificationRead
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Уведомление отмечено прочитанным
        default:
          $ref: '#/components/responses/Error'
  /api/notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Все уведомления отмечены прочитанными
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ToolId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [status, error, message]
      properties:
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        conflictToolId: { type: string, nullable: true }
        conflictToolName: { type: string, nullable: true }

    # ----- Auth -----
...
    MeResponse:
      type: object
      required: [id, username, fullName, role]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        fullName: { type: string }
        role: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/MeResponse' }
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          x-field-extra-annotation: '@field:Size(max = 255, message = "ФИО: до 255 символов")'
        username:
          type: string
          maxLength: 255
          x-field-extra-annotation: '@field:Pattern(regexp = "^[0-9]+$", message = "Логин Сигма должен содержать только цифры")'
        currentPassword: { type: string }
        newPassword:
          type: string
          x-field-extra-annotation: '@field:Size(min = 6, max = 128, message = "Пароль должен быть не короче 6 символов")'

    # ----- Tools -----
    ToolResponse:
      type: object
      required:
        - id
        - name
        - description
        - stage
        - status
        - roles
        - framework
        - constraints
        - ownerName
        - downloads
        - efficiencyPct
        - isTop
        - views
        - avgRating
        - ratingsCount
        - canManage
        - notesCount
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: string, nullable: true }
        stage: { type: string }
        status: { type: string }
        roles: { type: array, items: { type: string } }
        framework: { type: array, items: { type: string } }
        toolType: { type: string, nullable: true }
        constraints: { type: array, items: { type: string } }
        sourceLabel: { type: string, nullable: true }
        ownerName: { type: string }
        downloads: { type: integer, format: int32 }
        dau: { type: integer, format: int32, nullable: true }
        efficiencyPct: { type: integer, format: int32 }
        isTop: { type: boolean }
        views: { type: integer, format: int64 }
        avgRating: { type: number, format: double }
        ratingsCount: { type: integer, format: int32 }
        myRating: { type: integer, format: int32, nullable: true }
        canManage: { type: boolean }
        rejectionReason: { type: string, nullable: true }
        notesCount: { type: integer, format: int64 }
        segment: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RejectToolRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 1000
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank(message = "Укажите причину отклонения")'
    ArchiveToolRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 1000, nullable: true }
...
    UserResponse:
      type: object
      required: [id, username, fullName, role, enabled, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        fullName: { type: string }
        role: { type: string }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank'
    UpdateUserEnabledRequest:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }

    # ----- Admin: notes -----
    ToolNoteResponse:
      type: object
      required: [id, toolId, authorId, authorName, text, canManage, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        toolId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        authorName: { type: string }
        text: { type: string }
        canManage: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateToolNoteRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 2000
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank(message = "Введите текст заметки")'
    UpdateToolNoteRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 2000
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank(message = "Введите текст заметки")'

    # ----- Impact -----
    ImpactRowResponse:
      type: object
      required: [id, label, value, colorVariant, isFooter]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        value: { type: string }
        colorVariant: { type: string }
        isFooter: { type: boolean }
...
    UpdateImpactRowRequest:
      type: object
      required: [value]
      properties:
        value:
          type: string
          maxLength: 64
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotBlank'

    # ----- Notifications -----
    NotificationResponse:
      type: object
      required: [id, type, toolName, message, read, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        toolId: { type: string, format: uuid, nullable: true }
        toolName: { type: string }
        message: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
    UnreadCountResponse:
      type: object
      required: [count]
      properties:
        count: { type: integer, format: int64 }